+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -23,6 +23,9 @@ Notes](../../RELEASENOTES.md).
 * Added support for `Meter.TelemetrySchemaUrl` property.
   ([#6731](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6731))
 
+* Fix `NullReferenceException` when no bucket boundaries configured for a view.
+  ([#6773](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6773))
+
 ## 1.14.0
 
 Released 2025-Nov-12